+++ docs/history.md
@@ -17,6 +17,7 @@
 - Fixed the bug where the baby description would overlap with the streak text in certain situations. (Thanks Toooschi)
 - Fixed the bug where the custom baby timer was incorrectly placed in certain situations. (Thanks Gamonymous)
 - Fixed the bug where Turtle Dragon Baby would softlock with rooms that contained Flaming Hoppers. (Thanks SuperpaperChildy)
+- Fixed the bug where Beanie Baby could be assigned if the player did not have a trinket. (Thanks Gamonymous)
 
 ### _v1.7.1_ - September 8th, 2023